create components folder, move all components there
