+++ src/app/components/sidebar/sidebar.component.scss
